Move *Instance from package object to top-level files
