+++ hap-api/src/main/scala/com/github/otah/hap/api/CharacteristicInstance.scala
@@ -1,0 +1,10 @@
+package com.github.otah.hap.api
+
+import scalajson.ast.JObject
+
+import scala.concurrent.{ExecutionContext, Future}
+
+case class CharacteristicInstance(iid: InstanceId, characteristic: Option[LowLevelCharacteristic]) {
+
+  def asJson(implicit ec: ExecutionContext): Option[Future[JObject]] = characteristic map (_.asJson(iid.value))
+}
