+++ hap-api/src/main/scala/com/github/otah/hap/api/ServiceInstance.scala
@@ -1,0 +1,3 @@
+package com.github.otah.hap.api
+
+case class ServiceInstance(iid: InstanceId, service: AccessoryService)
