+++ hap-api/src/main/scala/com/github/otah/hap/api/package.scala
@@ -28,19 +28,4 @@ package object api {
 
     def jsonValueNotifier(implicit ec: ExecutionContext): Option[LowLevelNotifier]
   }
-
-  case class CharacteristicInstance(iid: InstanceId, characteristic: Option[LowLevelCharacteristic]) {
-
-    def asJson(implicit ec: ExecutionContext): Option[Future[JObject]] = characteristic map (_.asJson(iid.value))
-  }
-
-  case class ServiceInstance(iid: InstanceId, service: AccessoryService)
-
-  case class InstanceId(value: Int) {
-    require(value > 0, "IID has to be positive number")
-
-    def -(service: AccessoryService): ServiceInstance = ServiceInstance(this, service)
-    def -(ch: LowLevelCharacteristic): CharacteristicInstance = CharacteristicInstance(this, Some(ch))
-    def -(maybeCh: Option[LowLevelCharacteristic]): CharacteristicInstance = CharacteristicInstance(this, maybeCh)
-  }
 }
